# Add migrations for span and trace_meta tables