--- a/genai-engine/src/schemas/response_schemas.py
+++ b/genai-engine/src/schemas/response_schemas.py
@@ -604,6 +604,7 @@ class SpanResponse(BaseModel):
     span_id: str
     parent_span_id: Optional[str] = None
     span_kind: Optional[str] = None
+    span_name: Optional[str] = None
     start_time: datetime
     end_time: datetime
     task_id: Optional[str] = None
@@ -680,6 +681,7 @@ class SpanWithMetricsResponse(BaseModel):
     span_id: str
     parent_span_id: Optional[str] = None
     span_kind: Optional[str] = None
+    span_name: Optional[str] = None
     start_time: datetime
     end_time: datetime
     task_id: Optional[str] = None
@@ -705,6 +707,7 @@ class NestedSpanWithMetricsResponse(BaseModel):
     span_id: str
     parent_span_id: Optional[str] = None
     span_kind: Optional[str] = None
+    span_name: Optional[str] = None
     start_time: datetime
     end_time: datetime
     task_id: Optional[str] = None