… genai-engine/src/api_changelog.md
… genai-engine/src/services/span_query_service.py
… genai-engine/src/services/tree_building_service.py
… genai-engine/staging.openapi.json
… genai-engine/tests/routes/span/conftest.py
… genai-engine/tests/routes/span/test_metadata_management.py
… genai-engine/tests/routes/span/test_span_routes.py