# Database Optimization Results

## Tickets
[[GenAI Engine] Database Indexes, Query/Subquery Optimization for Performance](https://arthurai.atlassian.net/browse/UP-2984)
[[GenAI Engine] Extend Schema for Span Names](https://arthurai.atlassian.net/browse/UP-2985)

## Summary
**Strategic ingestion slowdown (+50-60% latency) delivers 17-32% faster queries and 43-46% reduced DB load**

## Key Optimizations
- **Pagination throughput**: +100-305% improvement by maintaining trace metadata instead of recomputing each time.
- **Preprocessing during ingestion**: Optimizes downstream query patterns

## Performance Comparison

### 10qpm Workload (9k-10k spans):
| Metric | Baseline | Optimized | Improvement |
|--------|----------|-----------|-------------|
| Query time per span | 0.000096s | 0.000080s | **17% faster** |
| Total records queried | 4,848 | 2,827 | **42% reduction** |
| Ingestion time per trace | 0.006s | 0.009s | **+50% latency** |

### 20qpm Workload (19k-20k spans):
| Metric | Baseline | Optimized | Improvement |
|--------|----------|-----------|-------------|
| Query time per span | 0.000148s | 0.000101s | **32% faster** |
| Total records queried | 14,818 | 7,828 | **47% reduction** |
| Ingestion time per trace | 0.005s | 0.008s | **+60% latency** |


## Impact
Excellent tradeoff for read-heavy workloads: accepts moderate ingestion cost to deliver substantial query performance gains and database efficiency improvements that scale with load.